// src/store/types.ts
export interface IState {
  token: string | null;
  pending: boolean;
  error: string | null;
  cart: CartItem[];
  orders: Order[];
  refs: Refs;
  favoriteProducts: Favorite[] | [];
  foundProducts: Product[];
  promoProducts: UserAction[];
  user: User;
  suggestions: {streets: any; houses: any};
  catalog: Category[] | [];
  products: {[key: string]: Product[]};
  reviews: Review[];
  exists: boolean;
  phoneNumber: string;
}

export type IAction = {
  data: any;
  type: string;
  error: string;
};

export interface OAction<Payload = {} | any> {
  type: string;
  payload: Payload;
}

export enum Action {
  // AUTH
  SIGN_UP_REQUEST = 'SIGN_UP_REQUEST',
  SIGN_UP_REJECTED = 'SIGN_UP_REJECTED',

  SIGN_IN_REQUEST = 'SIGN_IN_REQUEST',
  SIGN_IN_FULFILLED = 'SIGN_IN_FULFILLED',
  SIGN_IN_REJECTED = 'SIGN_IN_REJECTED',

  SIGN_IN_WITH_APPLE_REQUEST = 'SIGN_IN_WITH_APPLE_REQUEST',
  SIGN_IN_WITH_APPLE_REJECTED = 'SIGN_IN_WITH_APPLE_REJECTED',

  SIGN_IN_WITH_GOOGLE_REQUEST = 'SIGN_IN_WITH_GOOGLE_REQUEST',
  SIGN_IN_WITH_GOOGLE_REJECTED = 'SIGN_IN_WITH_GOOGLE_REJECTED',

  SIGN_OUT_REQUEST = 'SIGN_OUT_REQUEST',
  SIGN_OUT_FULFILLED = 'SIGN_OUT_FULFILLED',
  SIGN_OUT_REJECTED = 'SIGN_OUT_REJECTED',

  SIGN_IN_WITH_PHONE_NUMBER_REQUEST = 'SIGN_IN_WITH_PHONE_NUMBER_REQUEST',
  SIGN_IN_WITH_PHONE_NUMBER_REJECTED = 'SIGN_IN_WITH_PHONE_NUMBER_REJECTED',

  CHECK_PHONE_NUMBER_REQUEST = 'CHECK_PHONE_NUMBER_REQUEST',
  CHECK_PHONE_NUMBER_FULFILLED = 'CHECK_PHONE_NUMBER_FULFILLED',
  CHECK_PHONE_NUMBER_REJECTED = 'CHECK_PHONE_NUMBER_REJECTED',

  VERIFY_PHONE_NUMBER_REQUEST = 'VERIFY_PHONE_NUMBER_REQUEST',
  VERIFY_PHONE_NUMBER_FULFILLED = 'VERIFY_PHONE_NUMBER_FULFILLED',
  VERIFY_PHONE_NUMBER_REJECTED = 'VERIFY_PHONE_NUMBER_REJECTED',

  // CART
  FETCH_CART_REQUEST = 'FETCH_CART_REQUEST',
  FETCH_CART_FULFILLED = 'FETCH_CART_FULFILLED',
  FETCH_CART_REJECTED = 'FETCH_CART_REJECTED',
  ADD_TO_CART_REQUEST = 'ADD_TO_CART_REQUEST',
  ADD_TO_CART_REJECTED = 'ADD_TO_CART_REJECTED',
  REMOVE_FROM_CART_REQUEST = 'REMOVE_FROM_CART_REQUEST',
  REMOVE_FROM_CART_REJECTED = 'REMOVE_FROM_CART_REJECTED',
  UPDATE_CART_REQUEST = 'UPDATE_CART_REQUEST',
  UPDATE_CART_REJECTED = 'UPDATE_CART_REJECTED',
  UPDATE_LOCAL_CART_REQUEST = 'UPDATE_LOCAL_CART_REQUEST',

  // PRODUCTS
  FETCH_CATALOG_REQUEST = 'FETCH_CATALOG_REQUEST',
  FETCH_CATALOG_FULFILLED = 'FETCH_CATALOG_FULFILLED',
  FETCH_CATALOG_REJECTED = 'FETCH_CATALOG_REJECTED',
  FETCH_PRODUCTS_REQUEST = 'FETCH_PRODUCTS_REQUEST',
  FETCH_PRODUCTS_FULFILLED = 'FETCH_PRODUCTS_FULFILLED',
  FETCH_PRODUCTS_REJECTED = 'FETCH_PRODUCTS_REJECTED',
  FETCH_PROMO_PRODUCTS_REQUEST = 'FETCH_PROMO_PRODUCTS_REQUEST',
  FETCH_PROMO_PRODUCTS_FULFILLED = 'FETCH_PROMO_PRODUCTS_FULFILLED',
  FETCH_PROMO_PRODUCTS_REJECTED = 'FETCH_PROMO_PRODUCTS_REJECTED',
  SEARCH_PRODUCTS_REQUEST = 'SEARCH_PRODUCTS_REQUEST',
  SEARCH_PRODUCTS_FULFILLED = 'SEARCH_PRODUCTS_FULFILLED',
  SEARCH_PRODUCTS_REJECTED = 'SEARCH_PRODUCTS_REJECTED',
  SORT_PRODUCTS_BY_PRICE_REQUEST = 'SORT_PRODUCTS_BY_PRICE_REQUEST',

  // FAVORITE PRODUCTS
  FETCH_FAVORITES_REQUEST = 'FETCH_FAVORITES_REQUEST',
  FETCH_FAVORITES_FULFILLED = 'FETCH_FAVORITES_FULFILLED',
  FETCH_FAVORITES_REJECTED = 'FETCH_FAVORITES_REJECTED',
  ADD_TO_FAVORITES_REQUEST = 'ADD_TO_FAVORITES_REQUEST',
  ADD_TO_FAVORITES_REJECTED = 'ADD_TO_FAVORITES_REJECTED',
  REMOVE_FROM_FAVORITES_REQUEST = 'REMOVE_FROM_FAVORITES_REQUEST',
  REMOVE_FROM_FAVORITES_REJECTED = 'REMOVE_FROM_FAVORITES_REJECTED',
  UPDATE_LOCAL_FAVORITES_REQUEST = 'UPDATE_LOCAL_FAVORITES_REQUEST',

  // ORDERS
  FETCH_ORDERS_REQUEST = 'FETCH_ORDERS_REQUEST',
  FETCH_ORDERS_FULFILLED = 'FETCH_ORDERS_FULFILLED',
  FETCH_ORDERS_REJECTED = 'FETCH_ORDERS_REJECTED',
  ADD_CASH_ORDER_REQUEST = 'ADD_CASH_ORDER_REQUEST',
  ADD_CASH_ORDER_REJECTED = 'ADD_CASH_ORDER_REJECTED',
  ADD_BANK_ORDER_REQUEST = 'ADD_BANK_ORDER_REQUEST',
  ADD_BANK_ORDER_REJECTED = 'ADD_BANK_ORDER_REJECTED',

  // USER
  FETCH_REFS_REQUEST = 'FETCH_REFS_REQUEST',
  FETCH_REFS_FULFILLED = 'FETCH_REFS_FULFILLED',
  FETCH_REFS_REJECTED = 'FETCH_REFS_REJECTED',
  FETCH_USER_DATA_REQUEST = 'FETCH_USER_DATA_REQUEST',
  FETCH_USER_DATA_FULFILLED = 'FETCH_USER_DATA_FULFILLED',
  FETCH_USER_DATA_REJECTED = 'FETCH_USER_DATA_REJECTED',
  EDIT_USER_DATA_REQUEST = 'EDIT_USER_DATA_REQUEST',
  EDIT_USER_DATA_REJECTED = 'EDIT_USER_DATA_REJECTED',
  FETCH_REVIEWS_REQUEST = 'FETCH_REVIEWS_REQUEST',
  FETCH_REVIEWS_FULFILLED = 'FETCH_REVIEWS_FULFILLED',
  FETCH_REVIEWS_REJECTED = 'FETCH_REVIEWS_REJECTED',
  ADD_REVIEW_REQUEST = 'ADD_REVIEW_REQUEST',
  ADD_REVIEW_REJECTED = 'ADD_REVIEW_REJECTED',
  FETCH_STREETS_REQUEST = 'FETCH_STREETS_REQUEST',
  FETCH_STREETS_FULFILLED = 'FETCH_STREETS_FULFILLED',
  FETCH_STREETS_REJECTED = 'FETCH_STREETS_REJECTED',
  FETCH_HOUSES_REQUEST = 'FETCH_HOUSES_REQUEST',
  FETCH_HOUSES_FULFILLED = 'FETCH_HOUSES_FULFILLED',
  FETCH_HOUSES_REJECTED = 'FETCH_HOUSES_REJECTED',
}